+++ expo/gbe/scheduling/views/__init__.py
@@ -8,3 +8,4 @@ from class_wizard_view import ClassWizardView
 from manage_events_view import ManageEventsView
 from set_favorite_view import SetFavoriteView
 from copy_occurrence_view import CopyOccurrenceView
+from event_detail_view import EventDetailView
