+++ expo/gbe/templates/gbe/scheduling/calendar.tmpl
@@ -51,7 +51,10 @@
   {% endifchanged %}
     <div class="col-lg-{{grid_size}} col-md-{% if grid_size > 4 %}{{grid_size}}{% else %}4{% endif %} col-sm-{% if grid_size > 6 %}{{grid_size}}{% else %}6{% endif %} col-xs-12">
      {% if occurrence.favorite_link %}
-      <a href="{{occurrence.favorite_link}}?next={{ request.get_full_path }}" class="detail_link cal-favorite"><i class="fa {%if occurrence.highlight == "Interested" %}fa-star{% else %}fa-star-o{% endif %}" aria-hidden="true"></i></a>
+      <a href="{{occurrence.favorite_link}}?next={{ request.get_full_path }}"
+         class="detail_link cal-favorite" data-toggle="tooltip"
+         title="{%if occurrence.highlight == "Interested" %}Remove from Schedule{% else %}Add to Schedule{% endif %}">
+       <i class="fa {%if occurrence.highlight == "Interested" %}fa-star{% else %}fa-star-o{% endif %}" aria-hidden="true"></i></a>
      {% endif %}
      {{ occurrence.start }}-{{ occurrence.end }}<br>
      <a href="{{occurrence.detail_link}}" class="detail_link">{{ occurrence.title}}</a><br>
