+++ expo/scheduler/models.py
@@ -632,7 +632,7 @@ class Event(Schedulable):
         Return a detail link to self, with title as link text
         '''
         return '<a href="%s">%s</a>' % (reverse('detail_view',
-                                                urlconf='scheduler.urls',
+                                                urlconf='gbe.scheduling.urls',
                                                 args=[self.eventitem_id]),
                                         self.eventitem.describe)
