+++ expo/scheduler/views/__init__.py
@@ -1,8 +1,6 @@
 from remains import (
     get_events_display_info,
-    get_event_display_info,
     event_list,
-    detail_view,
     schedule_acts,
     delete_schedule,
     delete_event,
